Migrated from go-sdl2 to ebitengine. - Functionality wasn't changed - Unsurprisingly, memory leaks are gone

+++ config/config.go
@@ -6,5 +6,5 @@ const (
 )
 
 var (
-	FONT_SIZE int = 32
+	FONT_SIZE float64 = 32
 )
